requirements.txt
python-dotenv==1.0.0
google-generativeai==0.3.2
tweepy==4.14.0
python-dateutil==2.8.2